+++ src/gl_scene/generator.cpp
@@ -29,5 +29,5 @@ Generator::Generator() {
 }
 
 float Generator::Rand(float min, float max) {
-  return ((max-min)*((float)rand()/RAND_MAX))+min;
+  return ((max-min)*((float)rand()/(float)RAND_MAX))+min;
 }
